[UPDATE] Component GroupDialogAdd move to new folder
