+++ src/app/admin/group/entity.interface.ts
@@ -16,3 +16,7 @@ export interface Faculty {
   faculty_name: string;
   faculty_description: string;
 }
+
+export interface DialogData {
+    data: any;
+  }
