+++ src/app/http.interceptor.ts
@@ -5,7 +5,7 @@ import { Observable } from 'rxjs';
 @Injectable()
 export class ApiHttpInterceptor implements HttpInterceptor {
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    const apiReq = request.clone({ url: `http://dtapi.if.ua/${request.url}` });
+    const apiReq = request.clone({ url: `http://dtapi.if.ua/api/${request.url}` });
 
     return next.handle(apiReq);
   }
